requirements.txt
customtkinter
pdf2docx
Pillow
docx2pdf
